pyproject.toml
[tool.poetry]
name = "atomichabits"
version = "0.1.0"
description = ""
authors = ["u_alex90 <u_alex90@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
djangorestframework-simplejwt = "^5.3.0"
psycopg2-binary = "^2.9.9"
pillow = "^10.1.0"
django-filter = "^23.4"
django-environ = "^0.11.2"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
